use hover only prefetch link
